backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
anthropic
pytest
requests
supabase